+++ yaz_messaging_plugin/test/test_depth.py
@@ -11,7 +11,7 @@ triple_first.triple_second.triple_third: c
 """.lstrip()
     }
 
-    def test_010_cleanup(self):
+    def test_010_fix(self):
         expected = """
 double_first:
     double_second: b
@@ -22,10 +22,10 @@ triple_first:
 """.lstrip()
 
         caller = self.get_caller()
-        caller("cleanup", "--changes", "overwrite")
+        caller("fix", "--changes", "overwrite")
         self.assertEqual(expected, self.get_file_content("depth.nl.yml"))
 
-    def test_020_cleanup__max_depth_0(self):
+    def test_020_fix__max_depth_0(self):
         expected = """
 double_first.double_second: b
 single_first: a
@@ -33,10 +33,10 @@ triple_first.triple_second.triple_third: c
 """.lstrip()
 
         caller = self.get_caller()
-        caller("cleanup", "--changes", "overwrite", "--max-depth", "0")
+        caller("fix", "--changes", "overwrite", "--max-depth", "0")
         self.assertEqual(expected, self.get_file_content("depth.nl.yml"))
 
-    def test_030_cleanup__max_depth_1(self):
+    def test_030_fix__max_depth_1(self):
         expected = """
 double_first:
     double_second: b
@@ -46,10 +46,10 @@ triple_first:
 """.lstrip()
 
         caller = self.get_caller()
-        caller("cleanup", "--changes", "overwrite", "--max-depth", "1")
+        caller("fix", "--changes", "overwrite", "--max-depth", "1")
         self.assertEqual(expected, self.get_file_content("depth.nl.yml"))
 
-    def test_040_cleanup__max_depth_2(self):
+    def test_040_fix__max_depth_2(self):
         expected = """
 double_first:
     double_second: b
@@ -60,7 +60,7 @@ triple_first:
 """.lstrip()
 
         caller = self.get_caller()
-        caller("cleanup", "--changes", "overwrite", "--max-depth", "2")
+        caller("fix", "--changes", "overwrite", "--max-depth", "2")
         self.assertEqual(expected, self.get_file_content("depth.nl.yml"))
 
 
@@ -72,7 +72,7 @@ a.b.c: ABC
 """.lstrip()
     }
 
-    def test_010_cleanup__join(self):
+    def test_010_fix__join(self):
         expected = """
 a: A
 a.b:
@@ -80,10 +80,10 @@ a.b:
 """.lstrip()
 
         caller = self.get_caller()
-        caller("cleanup", "--changes", "overwrite", "--depth", "join")
+        caller("fix", "--changes", "overwrite", "--depth", "join")
         self.assertEqual(expected, self.get_file_content("depth.nl.yml"))
 
-    def test_020_cleanup__fail(self):
+    def test_020_fix__fail(self):
         caller = self.get_caller()
         with self.assertRaisesRegex(yaz.Error, r"Conflicting keys when expanding path .*"):
-            caller("cleanup", "--depth", "fail")
+            caller("fix", "--depth", "fail")