+++ yaz_messaging_plugin/test/test_google_translate.py
@@ -16,7 +16,7 @@ controller:
 """.lstrip()
     }
 
-    def test_010_cleanup__google_translate(self):
+    def test_010_fix__google_translate(self):
         expected_nl = """
 controller:
     action:
@@ -32,7 +32,35 @@ controller:
 """.lstrip()
 
         caller = self.get_caller()
-        caller("cleanup", "--changes", "overwrite", "--sync", "google-translate")
+        caller("fix", "--changes", "overwrite", "--sync", "google-translate")
+        self.assertEqual(expected_nl, self.get_file_content("translate.nl.yml"))
+        self.assertEqual(expected_en, self.get_file_content("translate.en.yml"))
+
+class TestGoogleTranslateWithEmptyString(TestCase):
+    files = {
+        "translate.nl.yml": """
+controller:
+    action:
+        empty: ''
+""".lstrip(),
+        "translate.en.yml": ""
+    }
+
+    def test_010_fix__google_translate(self):
+        expected_nl = """
+controller:
+    action:
+        empty: ''
+""".lstrip()
+
+        expected_en = """
+controller:
+    action:
+        empty: ''
+""".lstrip()
+
+        caller = self.get_caller()
+        caller("fix", "--changes", "overwrite", "--sync", "google-translate")
         self.assertEqual(expected_nl, self.get_file_content("translate.nl.yml"))
         self.assertEqual(expected_en, self.get_file_content("translate.en.yml"))
 
@@ -41,20 +69,24 @@ class TestGoogleTranslateWithPlaceholders(TestCase):
         "translate.nl.yml": """
 controller:
     action:
+        farewell_a: Tot ziens !username
+        farewell_b: Tot ziens !username, het was leuk dat je er was
         greeting_a: Hallo %username%
         greeting_b: Hallo %username%, het is vandaag %today% en je hebt nog %thingsTodo% dingen te doen
 """.lstrip(),
         "translate.en.yml": ""
     }
 
-    def test_010_cleanup__google_translate(self):
+    def test_010_fix__google_translate(self):
         expected_en = """
 controller:
     action:
+        farewell_a: Bye !username
+        farewell_b: Goodbye !username, it was nice that you had
         greeting_a: Hello %username%
         greeting_b: Hello %username%, today %today% and you have to do it %thingsTodo% things
 """.lstrip()
 
         caller = self.get_caller()
-        caller("cleanup", "--changes", "overwrite", "--sync", "google-translate")
+        caller("fix", "--changes", "overwrite", "--sync", "google-translate")
         self.assertEqual(expected_en, self.get_file_content("translate.en.yml"))