+++ CHANGELOG.md
@@ -22,3 +22,9 @@
 
 # 1.1.0
 - Use the `googletrans` library to translate missing values during sync
+
+# 1.1.1
+- Fix unit tests
+- Fix bug when trying to google translate an empty string
+- Add support for `!foo` placeholders using google translate
+- Show the diff when a change is detected (allowing someone to manually fix it)